Week 3: Interfaces and Polymorphism:
Step 1: I defined the Movable interface and applied it to Player and Box.

Step 2: I implemented the collision and movement logic (for walls and boxes) in Game.java.

Step 3: We just added the Scanner and game loop to Game.java to handle user input.

Step 4: This step is "Playing a few turns," which I can now do by running the Game.java file.